Added Save Settings For Wrap Panel Ratio

+++ Takvim/Converters/WrapPanelRatioConverter.cs
@@ -6,7 +6,7 @@ namespace Takvim
 {
     public class WrapPanelRatioConverter : IValueConverter
     {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture) => (value is double ratio) ? System.Windows.SystemParameters.WorkArea.Width / ratio : System.Windows.SystemParameters.WorkArea.Width / 2;
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture) => (value is short ratio) ? System.Windows.SystemParameters.WorkArea.Width / ratio : System.Windows.SystemParameters.WorkArea.Width / 2;
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotImplementedException();
     }
